rename diagnostic components
