+++ front-end/src/components/KeyMeasureDiagnostic/DiversificationMeasure.vue
