+++ front-end/src/components/KeyMeasureDiagnostic/InformationMeasure.vue
