+++ front-end/src/components/KeyMeasureDiagnostic/NoPlasticMeasure.vue
