+++ front-end/src/components/KeyMeasureDiagnostic/QualityMeasure.vue
