+++ front-end/src/components/KeyMeasureDiagnostic/WasteMeasure.vue
